requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
google-generativeai==0.3.2
pytest==8.0.0
pytest-cov==4.1.0
requests==2.31.0
python-multipart==0.0.9
loguru==0.7.2 